requirements.txt
Django==5.2.8
djangorestframework==3.15.0
django-cors-headers==4.0.0
pandas==2.2.3
openpyxl==3.1.2
gunicorn==20.1.0
whitenoise==6.5.0

# Optional (only if you integrate OpenAI or other libs)
openai==0.28.0
